requirements.txt
SpeechRecognition
pyttsx3
pyaudio
requests
python-dotenv